Initial commit for enabling pipeline phase tests
