+++ data_management/phase_tkinter_class.py
@@ -6,7 +6,15 @@ import cv2
 import tkinter as tk
 from tkinter import filedialog as fd
 import os
-
+import numpy as np
+import os.path as path
+import tkinter as tk
+import cv2
+import os
+import glob
+from enum import IntFlag, auto
+#from phase_tkinter_class import PipelinePhase
+#from phase_tkinter_class import np_photo_image
 
 def np_photo_image(image: np.ndarray):
     # This function creates the header information for PPM file format
@@ -22,6 +30,8 @@ def np_photo_image(image: np.ndarray):
 
 
 class PipelinePhase(tk.Frame):
+    import cv2
+    from enum import auto
     def __init__(self, next_phase, master=None, prev_phase: tk.Frame = None):
         super().__init__(master)
         self.next_phase = next_phase
@@ -39,7 +49,14 @@ class PipelinePhase(tk.Frame):
                 self.filename = os.path.join(*self.filename.split("/"))
                 self.filename = self.filename.replace(":", ":\\")
 
-            self.photo_folder = os.path.dirname(self.filename)
+            self.story_folder = os.path.dirname(self.filename)
+
+            # check if user opened a file in a "phase" folder
+            if "phase" in self.story_folder:
+                pass
+
+
+
             self.np_img = cv2.imread(self.filename, cv2.IMREAD_UNCHANGED - cv2.IMREAD_IGNORE_ORIENTATION)
 
             if len(self.np_img.shape) == 3:  # Color Image
@@ -49,7 +66,7 @@ class PipelinePhase(tk.Frame):
 
         else:
             self.np_img = prev_phase.np_img
-            self.photo_folder = prev_phase.photo_folder
+            self.story_folder = prev_phase.story_folder
             self.filename = prev_phase.filename
 
         self.photo_image = np_photo_image(self.np_img)
@@ -152,3 +169,15 @@ class PipelinePhase(tk.Frame):
             print("redraw to soon!")
         if len(self.canvas.children) > 0:
             self.redraw()
+
+    def save_button(self):
+        directory = path.dirname(self.filename)
+        filename, extension = path.basename(self.filename).split(".")
+        new_file_name = path.join(directory, self.phase, filename + "." + extension)
+        # convert before saving
+        self.np_img = np.array(cv2.cvtColor(self.np_img, cv2.COLOR_BGR2RGB))
+        cv2.imwrite(new_file_name, self.np_img)
+        # convert after saving so next phase gets correct image
+        self.np_img = np.array(cv2.cvtColor(self.np_img, cv2.COLOR_RGB2BGR))
+        self.filename = new_file_name
+        print(new_file_name)
